+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Missing modules AudioRecordManager, Speak, UPnPPortmap, and WorkDirectory.
+
 ## [0.1.0] - 2023-02-04
 
 ### Added